Filter OpenAPI `deepObject` parameters in Django
This provides a function `filter_deep_object_param`, which takes the
following parameters:

* `request` - the Django (or DRF) request object.
* `filter_prefix` - the name of the field being queried in the model.
* `param_name` - the parameter name for this filter (default: 'filter').
* `field_prefix` - any fields or relations that need to be prepended to the
  field name to reference the desired object.

It returns a Django `Q()` object to use in a `filter` (or `exclude` expression)
that filters on the fields and structures in the Django model.

As an example, this will match query parameters like:

1. `system_profile[sap_system]=true`
2. `system_profile[cpu_flags][contains]=clzero`
3. `system_profile[system_memory_bytes][gt]=4000000000`

And produce a filter equivalent to:

1. `system_profile__sap_system=True`
2. `system_profile__cpu_flags__contains='clzero'`
3. `system_profile__system_memory_bytes__gt=4000000000`

Note that for the `gt` operator the value is converted to a number, and
the special values '`true`', '`True`', '`false`' and '`False`' are
converted to Python `True` and `False` boolean values respectively.

Numbers do not match strings in JSON value introspection, so
`system_profile[number_of_sockets]=1` will convert to the filter
`system_profile__number_of_sockets='1'` and will not match a record where
number of sockets is the number `1`.  A value that is wholly digits will
be converted to an integer if the 'eq' or 'ne' operators are given, so
`system_profile[number_of_sockets][eq]=1` will match an integer number
of sockets value.

All these would be matched by an openapi.Parameter with the name
`system_profile`.  If more than one of these parameter constructions
appears in the parameters supplied, these will be ANDed together in the
returned filter.

The operators supported are based on:

https://github.com/RedHatInsights/insights-api-common-rails#usage

NOTE: also, because we don't attempt to bar any other comparator, because
they might also be a key in a dictionary, this also allows us to accept
all standard Django filter operators.  E.g. we accept both 'starts_with'
and 'startswith'.

Signed-off-by: Paul Wayper <paulway@redhat.com>